removed Plastic Source Control and modified prefab constraints and physic material
I looked at your project and I understand why the issue has been confounding. You had everything set up properly, but for some reason the rigidbody was not acknowledging that a material was applied to it.

I ran the project and saw the ball land and continue to roll.

I looked at the projectile in the scene view and used shift-F to follow it. It was rotating about the Z axis despite having a rotation constraint.

First I tried attaching the physic material to the ground to try to get some kind of response - nothing appreciable. Weird result.

I disconnected the physic material from the prefab and reset the constraints. Then reset the slingshot prefab to projectile to see if it would force it to recognize the new settings.

The ball landed and slid, It did not roll.

I reconnected the physic material - it bounced really high (ground plus projectile bounciness)

Removed the physic material from the ground.

Projectile behaves as expected.